+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.1.0] - 2025-02-?
 
+### Added
+
+- Add config option to make totem infinity if enchanted with Infinity effect
+
 ### Changed
 
 - Can add custom effects from UI